Return only gases with prices
The name of the function implies that it is returning
only gases which have pricing. So it shoult NOT
return gases without price. Althought there should never
be a gas without price...

At the time that test data includes gases without price anyway.
This is going to be fixed once default prices are applied in
becoming migration.